% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hwlocExample}
\alias{hwlocExample}
\title{Example hwloc Function}
\usage{
hwlocExample()
}
\value{
Nothing is returned, the function is invoked for the side effect of
displaying to the standard output.
}
\description{
This function provides a simple introduction and overview of \code{hwloc}
by printing some diagnostics to the standard output
}
