\name{RcppHwloc-package}
\alias{RcppHwloc-package}
\alias{RcppHwloc}
\docType{package}
\title{\packageTitle{RcppHwloc}}
\description{\packageDescription{RcppHwloc}}
\section{Package Content}{\packageIndices{RcppHwloc}}
\author{\packageAuthor{RcppHwloc}}
\section{Maintainer}{\packageMaintainer{RcppHwloc}}
\keyword{package}
